Move MyRFQ into new file as per TODO
