+++ src/lib.rs
@@ -1,10 +1,12 @@
 pub mod bindings;
 pub mod configuration;
 pub mod indexer;
-pub mod request_for_quote {
-    tonic::include_proto!("rfq");
-}
 pub mod routes;
+pub mod services;
 pub mod startup;
 pub mod structs;
 pub mod telemetry;
+
+pub mod request_for_quote {
+    tonic::include_proto!("rfq");
+}
